Superstore Analysis
Identifying profits in each sub-category per region and profits, also profits per product id and ship-mode. The purpose of the project was to use a data visualization tool to answer the following questions: Which are the two biggest profit centers and two biggest loss-makers? Which products should the superstore stop selling? Which product subcategories should the store focus on and which should they stop selling? Which states should the Superstore consider discontinuing business in based on Negative Profit visualization?