Add files via upload
The executive team is interested in understanding how well the website is converting product page views into purchases. This project challenge your skills in cleaning raw data and producing a statistical analysis. The first step was to create cohorts, using the raw data, using the raw data, putting users together by years. Then was to analyze that data to highlight the conversion rate from impressions to purchases. Then analyzing that data to produce the retention rate of users from year to year. This anaylsis required the usage of multiple business analytics skills.