https://public.tableau.com/views/First_Dashboard_17235236892570/UnderstandingProfitsBySub-Category?:language=en-US&publish=yes&:sid=&:redirect=auth&:display_count=n&:origin=viz_share_link
https://public.tableau.com/views/First_Dashboard_17235236892570/UnderstandingProfitsBySub-Category?:language=en-US&publish=yes&:sid=&:redirect=auth&:display_count=n&:origin=viz_share_link
The superstore is in peril! You have been hired as a consultant to review the superstore’s operations and increase its profitability to avoid bankruptcy. This project utilized Tableau visualization tool. I started by identifying the biggest profit makers and the biggest loss makers. This will allow the Superstore to make adjustments to products that great for business and those that the superstore should suspend sells. Then an analysis for advertising, I used the states and months of the year. The profit margin for this analysis was 20%. The return on advertising was above 20 %. The last part of this analysis was to highlight the products with high return rates. Return rates reduce profit margins. Products with high return rates effect profit margins, locating these states and these products will allow skateholders to make decision on theses indicators.